dmeventd: ensure service gets stopped normally on shutdown
cherry-pick two patches which ensure that dmeventd gets always
stopped when requested (i.e. if exiting is not ok at the moment it
gets scheduled for the next possible moment and not thrown away) and
which also ensures that it gets stopped more safely on shutdown.

This addresses a issue where when LVM(thin) was used on a host
shutdown could hang need 90 seconds longer, seemingly hanging for
this time as without debug log active no message got shown.